using GeoPhyInv
using Test
using SparseArrays

Intro

GeoPhyInv.AGeom — Type

AGeom=Vector{AGeomss,1}

An acquisition geometry, bundled into the mutable AGeom type, has to be specified in order to create any Expt variable. Acquisition has supersources, sources and receivers. For AGeom is a mutable type to define source-receiver geometry. Each supersource has ns sources that are injected (or active) simultaneously, and a set of nr receivers to record. AGeomss is a subtype for each supersource, therefore: This package provides tools to easily create commonly-used AGeom instances, after deciding on a spatial grid mgrid for the experiment.

Indexing and Fields

Let ageom be an instance of this type, then fields can be accessed using:

ageom[i] : acquisition for ith supersource

ageom[i].sx : x positions of sources

ageom[i].sz : z positions of sources

ageom[i].rx : x positions of receivers

ageom[i].rz : z positions of receivers

ageom[i].ns : number of sources

ageom[i].nr : number of receivers

ageom=AGeom(mgrid, attrib, SSrcs(10), Recs(20))

Return pre-defined acquisition geometries (with 10 sources and 20 receivers), based on attrib, on an input mesh mgrid. The sources and receivers positions are same for all supersources, and are not placed on the edges of the mesh. Choose attrib::Symbol

=:xwell cross-well acquisition;

=:surf surface acquisition;

=:vsp vertical seismic profiling;

=:rvsp reverse vertical seismic profiling;

=:downhole downhole sources and receivers.

Examples

Lets create a mgrid for the experiment.

mgrid = [range(0.0, stop=10.,step=0.1), range(0.0, stop=30.,step=0.2)];

Then initialize an acquisition on mgrid.

ageom=AGeom(mgrid, SSrcs(10), Srcs(10), Recs(10));

Otherwise, initialize with one of the predefined acquisitions.

ageom=AGeom(mgrid, :xwell, SSrcs(10), Recs(10));

By default, the sources and receivers are randomly placed on the grid. If unsure, test it.

@test (ageom ∈ mgrid)

Test Passed

The source and receiver positions can be updated as desired.

update!(ageom[1], Srcs(), [0,1], [10,20],);
update!(ageom[1], Recs(), [0,0], [10,20],);
update!(ageom, SSrcs(), [0,1], [10,20], );

It is easy to combine supersources. Now ageom2 has 20 supersources.

ageom2=vcat(ageom, ageom);

Methods

GeoPhyInv.update! — Method

update!(ageom[1], Srcs(10), p1, p2)

Update source geometry for the first supersource, with 10 sources placed regularly between p1=[z1,x1] and p2=[z2,x2]

update!(ageom[2], Srcs(10), p0, rad, angles)

Update source geometry for the second supersource, with 10 sources placed equidistant from p0=[z0,x0] given radius rad and angles angles=[0,2pi].

GeoPhyInv.update! — Method

update!(ageom[1], Recs(10), args...)

Similar to updating source positions, but for receivers.

GeoPhyInv.update! — Method

update!(ageom, SSrcs(), p1, p2)
update!(ageom, SSrcs(), p0, rad, angles)

Update source positions for all supersources.

GeoPhyInv.update! — Method

update!(ageom, Recs(), p1, p2)
update!(ageom, Recs(), p0, rad, angles)

Update receiver positions for all supersources.

Base.in — Method

ageom ∈ mgrid
in(ageom, mgrid)

Assert if all the sources and receivers in AGeom are within bounds of mgrid.

Base.isequal — Method

isequal(ageom1, ageom2)

Assert if ageom1 equals ageom2.

Missing docstring.

Missing docstring for SparseArrays.SparseMatrixCSC(::AGeomss,::AbstractVector{StepRangeLen}). Check Documenter's build log for details.