This page was generated on 2021-11-29

using GeoPhyInv
using Test
using Random

Intro

GeoPhyInv.Records — Type

A mutable type that bundles multi-component records at receiver array.

records=Records(tgrid, ageom, [:p, :vx])

Here, we initialized records, of :p and :vx fields, in time domain for receivers and supersources in ageom.

Indexing

records[i] : the records due to ith supersource of all fields

records[i][:p] : extract field :p

records[i][:nr] : number of receivers (same as in ageom)

records.grid : returns tgrid

As mutable objects in Julia are like containers that might hold different values over time, records can be modified.

Examples

Define an acquisition geometry.

mgrid = [range(0.0, stop=10.,step=0.1), range(0.0, stop=30.,step=0.2)];
ageom=AGeom(mgrid, SSrcs(10), Srcs(10), Recs(10));

Need a time grid.

tgrid=range(0, stop=1.0, step=0.1);

Lets initialize records for :P and :vz fields.

data=Records(tgrid, ageom, [:p,:vz]);

Fill the :P field of 3rd supersource with random numbers.

Random.randn!(data[3][:p]);

Methods

Methods listed for SrcWav can used on instances of Records too.