using BenchmarkTools
using GeoPhyInv
using Test

Intro

GeoPhyInv.SeisInvExpt — Function

pa=SeisInvExpt(attrib_mod, attrib_inv; srcwav, ageom, tgrid, mediumm, mediumm_obs, dobs)

Arguments

attrib_mod::Union{Fdtd,FdtdBorn} : choose modeling attribute

attrib_inv::Union{LS,LS_prior,Migr,Migr_FD} : choose inversion attribute

obj::Union{LS,LS_prior} : which objective function?

=LS() : least-squares inversion

=LS_prior([1.0, 0.5]) regularized least-squares inversion

=Migr() : migration image

=Migr_FD() : finite-difference gradient (used for testing only)

Keyword Arguments

srcwav::SrcWav : source wavelets

ageom::AGeom : acquisition

tgrid::StepRangeLen : modelling time grid

mediumm::Medium : an instance of Medium (used to generate initial model on the inversion grid)

dobs::Data : observed data (required only when attrib==:field)

mediumm_obs::Medium : medium used for generating observed data (required only when attrib==:synthetic)

Optional Keyword Arguments

igrid : inversion grid (defaults to modeling grid)

igrid_interp_scheme : interpolation scheme

=:B1 linear

=:B2 second order

mediumm0 : background seismic model for Born modelling and inversion

parameterization::Vector{Symbol} : subsurface parameterization

=[:χvp, :χrho, :null] for vp and density contrast inversion

=[:null, :χrho, :null] for density-only contrast inversion

=[:χvp, :null, :null] for vp-only contrast inversion

tgrid_obs::StepRangeLen : time grid for observed data

attrib::Symbol=:synthetic : an attribute to control class

=:synthetic synthetic data inversion

=:field field data inversion

pa=SeisInvExpt(attrib_mod,attrib_inv, attrib; rfields, parameterization)

Predefined gallery of SeisInvExpt. Choose attrib::Symbol

=:pizza an experiment fast enough to be run on a laptop

=:downhole sources and receivers on a drill-string

Base.getindex — Method

Indexing

pa is an instance of SeisForwExpt.

pa[:data,i] : get data for ith supersource (defaults to first), after update!

pa[:ageom] : get ageom originally input while creating pa

pa[:srcwav]: srcwav used to create pa

pa[:snaps,i] : snapshots corresponding to tsnaps and ith source, after update!

Methods

GeoPhyInv.update! — Method

update!(pa)

Full Waveform Inversion using Optim and Ipopt packages. This method updates pa.mediumm and pa.dcal. More details about the optional parameters can be found in the documentation of the Optim package. The gradient is computed using the adjoint state method. pa.mediumi is used as initial model if non-zero.

Arguments

pa::PFWI : inversion object (updated inside method)

Keywords Arguments

solver Choose Symbol

=:ipopt for Ipopt.jl

=:optim for Optim.jl

Optional Arguments

optim_options : see Optim.jl package for optimization options

optim_scheme=LBFGS() : see Optim.jl for other options

bounded_flag=true : constrained inversion using medium parameter bounds (Ipopt.jl only has bounded inversion)

ipopt_options : see Ipopt.jl

update!(pa)

Return gradient at the first iteration, i.e., a migration image, when attrib_inv==Migr().

update!(pa)

Return gradient at the first iteration, i.e., a migration image, without using adjoint state method, when attrib_inv==Migr_FD().

LinearMaps.LinearMap — Method

F=LinearMap(pa)

If pa is an instance of SeisInvExpt, then return the linearized forward modeling operator F, such that F*x can be computed without explicitly storing the operator matrix (see LinearMaps.jl). The imaging/migration operator is given by transpose(F). These operators are the building blocks of iterative optimization schemes.