using GeoPhyInv
using LinearAlgebra
using Random

Intro

GeoPhyInv.SrcWav — Type

A mutable type that bundles multi-component source wavelets.

srcwav=SrcWav(tgrid, ageom, [:P, :Vx])

Here, we initialized wavelets, for :P and :Vx fields, in time domain for sources and supersources in ageom.

Indexing

srcwav[i] : wavelets for ith supersource of all fields

srcwav[i][:P] : extract field :P

srcwav[i][:ns] : number of sources (same as in ageom)

srcwav.grid : returns tgrid

As mutable objects in Julia are like containers that might hold different values over time, srcwav can be modified.

Examples

Define an acquisition geometry.

mgrid = [range(0.0, stop=10.,step=0.1), range(0.0, stop=30.,step=0.2)];
ageom=AGeom(mgrid, SSrcs(10), Srcs(10), Recs(10));

Need a time grid.

tgrid=range(0, stop=1.0, step=0.1);

Lets initialize records for :P field.

srcwav=SrcWav(tgrid, ageom, [:P]);

Fill the :P field of 3rd supersource with random numbers.

Random.randn!(srcwav[3][:P]);

Often we want to populate the same source wavelet to all the supersources and sources.

x=randn(length(tgrid));
update!(srcwav, [:P,], x);

Populate two different wavelets for first and second supersources.

x1=randn(length(tgrid));
x2=randn(length(tgrid));
update!(srcwav[1], [:P,], x1);
update!(srcwav[2], [:P,], x2);

Scale srcwav by a scalar overwriting it in-place.

rmul!(srcwav, 2.0);

Methods

Most of the methods listed below are also applicable to individual elements of srcwav.

Missing docstring.

Missing docstring for update!(GeoPhyInv.VNamedD, ::Vector{Symbol}, ::AbstractVector{Float64},). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.reverse!(::GeoPhyInv.VNamedD). Check Documenter's build log for details.

Base.iszero — Method

iszero(srcwav)

Returns bool if srcwav has all zeros.

Base.isequal — Method

isequal(srcwav1, srcwav2)

Assert if srcwav1 and srcwav2 are equal.

Missing docstring.

Missing docstring for issimilar(::GeoPhyInv.VNamedD, ::GeoPhyInv.VNamedD). Check Documenter's build log for details.

Base.vec — Method

v=vec(srcwav)
Reshape `srcwav` as a one-dimensional column vector.

Missing docstring.

Missing docstring for Random.randn!(::GeoPhyInv.VNamedD). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.fill!(::GeoPhyInv.VNamedD, ::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinearAlgebra.dot(::GeoPhyInv.VNamedD, ::GeoPhyInv.VNamedD). Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinearAlgebra.rmul!(::GeoPhyInv.VNamedD, ::Number). Check Documenter's build log for details.

Base.copyto! — Method

copyto!(srcwav1, srcwav2)

Copy srcwav2 to srcwav1, which has same size.

Base.copyto! — Method

copyto!(v, srcwav)

Same as vec(srcwav), but in-place operation.

Base.copyto! — Method

v=vec(srcwav) 
randn!(v)
copyto!(srcwav, v)

Copy v to srcwav. The sizes of the two objects much match. No memory allocations.