This page was generated on 2021-11-09

using GeoPhyInv
using Test

Intro

GeoPhyInv.AGeom — Type

AGeom=Vector{AGeomss,1}

An acquisition geometry, bundled into the mutable AGeom type, has to be specified in order to create any Expt variable. Acquisition has supersources, sources and receivers. For AGeom is a mutable type to define source-receiver geometry. Each supersource has ns sources that are injected (or active) simultaneously, and a set of nr receivers to record. AGeomss is a subtype for each supersource, therefore: This package provides tools to easily create commonly-used AGeom instances, after deciding on a spatial grid mgrid for the experiment.

Indexing and Fields

Let ageom be an instance of this type, then fields can be accessed using:

ageom[i] : acquisition for ith supersource

ageom[i].s[p] : position coordinates of sources where p ∈ [:z, :x, :y]

ageom[i].r[p] : position coordinates of receivers

ageom[i].ns : number of sources

ageom[i].nr : number of receivers

ageom=AGeom(mgrid, attrib, SSrcs(10), Recs(20))

Return pre-defined acquisition geometries (in this case with 10 supersources and 20 receivers), based on attrib, on an input mesh mgrid. The sources and receivers positions are same for all supersources, and are not placed on the edges of the mesh. Choose attrib::Symbol

=:xwell cross-well acquisition;

=:surf surface acquisition;

=:vsp vertical seismic profiling;

=:rvsp reverse vertical seismic profiling;

=:downhole downhole sources and receivers.

TODO: For 3D, just mean(ygrid) is used, need to work on more exotic gallery of 3D acquisitions.

Examples

Lets create a 2-D mgrid for the experiment.

mgrid2 = [range(0.0, stop = 10.0, step = 0.1), range(0.0, stop = 30.0, step = 0.2)];

Then initialize an acquisition on mgrid, where the positions will be randomly chosen.

ageom2 = AGeom(mgrid2, SSrcs(10), Srcs(10), Recs(10));

Similarly, we can do it for 3D too.

mgrid3 = fill(range(-10, stop=10, step=0.01),3);
ageom3 = AGeom(mgrid3, SSrcs(10), Srcs(10), Recs(10));

You can check the number of dimensions.

ndims(ageom2), ndims(ageom3) == 2, 3

(2, false, 3)

For 2D, we can also initialize with one of the predefined acquisitions.

ageom2 = AGeom(mgrid2, :xwell, SSrcs(10), Recs(10));

By default, the sources and receivers are randomly placed on the grid. If unsure, test it.

@test (ageom2 ∈ mgrid2)
@test (ageom3 ∈ mgrid3)

Test Passed

The source and receiver positions can be updated as desired.

update!(ageom2[1], Srcs(), [0, 1], [10, 20]);
update!(ageom2[1], Recs(), [0, 0], [10, 20]);
update!(ageom2, SSrcs(), [0, 1], [10, 20]);

It is easy to combine supersources. Now ageom2 has 20 supersources.

ageom2_new = vcat(ageom2, ageom2);

Methods

GeoPhyInv.update! — Method

update!(ageom[1], Srcs(10), p1, p2)

Update source geometry for the first supersource, with 10 sources placed regularly between p1=[z1,x1] and p2=[z2,x2]

update!(ageom[2], Srcs(10), p0, rad, angles)

Update source geometry for the second supersource, with 10 sources placed equidistant from p0=[z0,x0] given radius rad and angles angles=[0,2pi].

GeoPhyInv.update! — Method

update!(ageom[1], Recs(10), args...)

Similar to updating source positions, but for receivers.

GeoPhyInv.update! — Method

update!(ageom, SSrcs(), p1, p2)
update!(ageom, SSrcs(), p0, rad, angles)

Update all source positions in each supersources.

GeoPhyInv.update! — Method

update!(ageom, Recs(), p1, p2)
update!(ageom, Recs(), p0, rad, angles)

Update receiver positions for all supersources.

Base.in — Method

ageom ∈ mgrid
in(ageom, mgrid)

Assert if all the sources and receivers in AGeom are within bounds of mgrid.

Base.isequal — Method

isequal(ageom1, ageom2)

Assert if ageom1 equals ageom2.

SparseArrays.SparseMatrixCSC — Method

A=SparseArrays.SparseMatrixCSC(ageom[1],mgrid)
D=A*P

Return a sparse matrix that restricts the field P on mgrid to receiver positions, to give D.