using GeoPhyInv
using Test
using Random

Intro

GeoPhyInv.Data — Type

A mutable type that bundles multi-component records at multiple receivers.

data=Data(tgrid, ageom, [:P, :Vx])

Here, we initialized records, of :P and :Vx fields, in time domain for receivers and supersources in ageom.

Indexing

data[i] : the records due to ith supersource of all fields

data[i][:P] : extract field :P

data[i][:nr] : number of receivers (same as in ageom)

data.grid : returns tgrid

As mutable objects in Julia are like containers that might hold different values over time, data can be modified.

Examples

Define an acquisition geometry.

mgrid = [range(0.0, stop=10.,step=0.1), range(0.0, stop=30.,step=0.2)];
ageom=AGeom(mgrid, SSrcs(10), Srcs(10), Recs(10));

Need a time grid.

tgrid=range(0, stop=1.0, step=0.1);

Lets initialize records for :P and :Vz fields.

data=Data(tgrid, ageom, [:P,:Vz]);

Fill the :P field of 3rd supersource with random numbers.

Random.randn!(data[3][:P]);

Methods

Methods listed for SrcWav can used on instances of Data too.