This page was generated on 2021-11-15

using BenchmarkTools
using GeoPhyInv
using Test

Intro

GeoPhyInv.SeisForwExpt — Function

pa = SeisForwExpt(attrib_mod; ageom, srcwav, medium, tgrid);

Method to create an instance of SeisForwExpt. The output of this method can be used as an input to the in-place method update!, to actually perform a finite-difference modeling.

Arguments

attrib_mod : attribute to choose the type of modeling. Choose from

=FdtdAcoustic() for acoustic modeling

=FdtdElastic() for solving the isotropic elastic wave equation

=FdtdAcousticBorn() for Born modeling

Keyword Arguments

medium : medium parameters

tgrid : modeling time grid, maximum time in tgrid should be greater than or equal to maximum source time, same sampling interval as in srcwav

ageom : acquisition

srcwav : source wavelets

Optional Keyword Arguments

sflags=2 : source related flags

=0 inactive sources

=1 sources with injection rate

=2 volume injection sources

=3 sources input after time reversal (use only during backpropagation)

rflags=1 : receiver related flags

=0 receivers do not record (or) inactive receivers

=1 receivers are active only for the second propagating wavefield

rfields=[:p] : multi-component receiver flag. Choose Vector{Symbol}

=[:p] record pressure

=[:vx] record horizontal component of particle velocity

=[:vz] record vertical component of particle velocity

=[:p, :vx] record both pressure and velocity

tsnaps : store snapshots at these modeling times (defaults to half time)

=[0.1,0.2,0.3] record at these instances of tgrid

snaps_field::Symbol=:tauxx : which field to record?

zfree : z-coordinate of a stress-free horizontal layer

verbose::Bool=false : verbose flag

Base.getindex — Method

Indexing

pa is an instance of SeisForwExpt.

pa[:data,i] : get data for ith supersource (defaults to first), after update!

pa[:ageom] : get ageom originally input while creating pa

pa[:srcwav]: srcwav used to create pa

pa[:snaps,i] : snapshots corresponding to tsnaps and ith source, after update!

Methods

GeoPhyInv.update! — Method

update!(pa)

In-place method to perform the experiment and update pa after wave propagation. After update, see pa[:data] and pa[:snaps].

GeoPhyInv.update! — Method

update!(pa,medium_new)

Update pa with a new bundle of medium parameters medium_new, without additional memory allocation. This routine is used during inversion, where medium parameters are iteratively updated. The ability to iteratively run the forward mediuming task (with no additional memory allocation) on various subsurface mediums is necessary while implementing inversion algorithms.

GeoPhyInv.update! — Method

update!(pa,srcwav_new,sflags)

Update pa with a new bundle of source wavelets srcwav_new, without additional memory allocation. Optionally, sflags can be changed.