using BenchmarkTools
using GeoPhyInv
using Test
using Plots
gr()

Plots.GRBackend()

Intro

GeoPhyInv.Medium — Type

Mutable type for storing medium parameters.

mod=Medium(mgrid, names=[:vp,:vs,:rho]) # initiate elastic medium
mod=Medium(mgrid, names=[:vp,:rho]) # initiate acoustic medium

This initializes a subsurface model with :vp, :rho and :vs parameters. Print the names of the medium parameters stored in mod.

names(mod)

Indexing

mod.mgrid : returns the spatial-grid bundle

mod[:vp] : P-wave velocity

mod[:vs] : S-wave velocity

mod[:rho] : mass density

mod[:Zp] : P-wave impedance

mod[:K] : bulk modulus (unrelaxed when considering attenuation)

mod[:M] : P-wave modulus

mod[:mu] : shear modulus

mod[:Q] : quality factor (relaxation times are optimized to be constant over all frequencies; see Robertsson, et. al, 1994)

mod.ref : reference medium parameters

mod.bounds : bounds of medium parameters

GeoPhyInv.Medium — Method

mod=Medium(attrib)

As of now, only seismic models are predefined in this package. Choose attrib::Symbol

=:acou_homo2D : a test homogeneous acoustic model

=:acou_homo2 : a test homogeneous acoustic model, but with coarser spatial sampling (faster testing)

=:marmousi2 : marmousi2 model with lower resolution; useful for surface seismic experiments

=:marmousi2_small : a smaller section of marmousi2

=:pizza : :acou_homo2 with some perturbations

Examples

Load a predefined model.

medium = Medium(:elastic_homo3D);

Get the medium parameters that are stored.

names(medium)

1-element Vector{Vector{Symbol}}:
 [:vp, :vs, :rho]

Look at the reference values.

medium.ref

8-element Named Vector{Float64}
vcat  │ 
──────┼────────────
:vp   │      2500.0
:vs   │      1500.0
:rho  │      2500.0
:K    │  1.27917e10
:KI   │  3.8429e-10
:rhoI │  0.00047619
:mu   │      7.75e9
:muI  │ 3.69048e-10

Inspect the bounds.

medium.bounds

8-element Named Vector{Vector{Float64}}
vcat  │ 
──────┼───────────────────────────
:vp   │           [1500.0, 3500.0]
:vs   │           [1000.0, 2000.0]
:rho  │           [1500.0, 3500.0]
:K    │      [1.375e9, 2.42083e10]
:KI   │ [4.13081e-11, 7.27273e-10]
:rhoI │ [0.000285714, 0.000666667]
:mu   │            [1.5e9, 1.4e10]
:muI  │ [7.14286e-11, 6.66667e-10]

Plotting is easy

p1=heatmap(medium, :vp)
plot(p1,size=(800,400))

To construct an instance of 2-D Medium, we need ranges for z and x.

mgrid = [range(0.0, stop = 10.0, step = 0.1), range(0.0, stop = 30.0, step = 0.2)];

For 3-D, it is z, y and x.

mgrid = fill(range(-10, stop = 10.0, step = 0.1), 3)

3-element Vector{StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}}}:
 -10.0:0.1:10.0
 -10.0:0.1:10.0
 -10.0:0.1:10.0

Allocate basic medium parameters on the grid.

medium = Medium(mgrid, [:vp, :rho, :vs]);

Bounds for these parameters should be input for modeling or inversion. Use update!

vpb = [2100.0, 2200.0];
vsb = [1500, 1700];
rhob = [2100.0, 2300.0];
update!(medium, [:vp, :vs, :rho], [vpb, vsb, rhob]);

Just fill medium with average (reference) values.

fill!(medium);

Once the basic medium parameters are input, we can access some other derived parameters.

medium[:Zp];

Otherwise, we can manually update parameters of medium.

medium[:vp] .= 3000.0;
medium[:vs] .= 2000.0;

println(medium)

> 3-D Medium with [:z, :y, :x] in [[-10.0, 10.0], [-10.0, 10.0], [-10.0, 10.0]]
> number of samples:	[201, 201, 201]
> sampling intervals:	[0.1, 0.1, 0.1]
> vp:	min	3000.0	max	3000.0
> vs:	min	2000.0	max	2000.0
> rho:	min	2200.0	max	2200.0
> Bounds:
8-element Named Vector{Vector{Float64}}
vcat  │
──────┼───────────────────────────
:vp   │           [2100.0, 2200.0]
:rho  │           [2100.0, 2300.0]
:vs   │           [1500.0, 1700.0]
:K    │       [2.961e9, 2.26933e9]
:KI   │ [4.40658e-10, 3.37724e-10]
:rhoI │  [0.000434783, 0.00047619]
:mu   │         [4.725e9, 6.647e9]
:muI  │  [1.50444e-10, 2.1164e-10]

A model can be also be updated by adding random noise.

update!(medium, [:vp, :rho], randn_perc = 1.0);

Methods

Missing docstring.

Missing docstring for `$@example doc md # Base.getindex(::Medium, ::Symbol)$

GeoPhyInv.update!(::GeoPhyInv.Medium, ::Vector{Symbol},) Base.copyto!(x::AbstractArray, medium::Medium, fields::Vector{Symbol}) Base.vec(medium::Medium, ::Symbol)`. Check Documenter's build log for details.