using BenchmarkTools
using GeoPhyInv
using Test
using Plots
gr()

Intro

GeoPhyInv.Medium — Type

Mutable type for storing medium parameters.

mod=Medium(mgrid, names=[:vp,:rho])

This initializes a subsurface model with :vp and :rho parameters. Print the names of the medium parameters stored in mod.

names(mod)

Indexing

mod.mgrid : returns the spatial-grid bundle

mod[:vp] : P-wave velocity

mod[:rho] : mass density

mod[:Zp] : P-wave impedance

mod[:K] : bulk modulus

mod.ref : reference medium parameters

mod.bounds : bounds of medium parameters

GeoPhyInv.Medium — Method

mod=Medium(attrib)

As of now, only seismic models are predefined in this package. Choose attrib::Symbol

=:acou_homo1 : a test homogeneous acoustic model

=:acou_homo2 : a test homogeneous acoustic model, but with coarser spatial sampling (faster testing)

=:marmousi2 : marmousi2 model with lower resolution; useful for surface seismic experiments

=:pizza : :acous_homo2 with some perturbations

Examples

Load a predefined model.

mod=Medium(:marmousi2);
nothing #hide

Get the medium parameters that are stored.

names(mod)

Look at the reference values.

mod.ref

Inspect the bounds.

mod.bounds

Plotting is easy

p1=heatmap(mod, :vp)
plot(p1,size=(800,400))

To construct an instance of Medium, we need a 2-D grid.

mgrid = [range(0.0, stop=10.,step=0.1), range(0.0, stop=30.,step=0.2)];

Allocate basic medium parameters on the grid.

mod = Medium(mgrid, [:vp,:rho,:vs]);
nothing #hide

Bounds for these parameters should be input for modeling or inversion. Use update!

vpb=[2100.,2200.]; vsb=[1500, 1700]; rhob=[2100., 2300.];
update!(mod, [:vp,:vs,:rho], [vpb, vsb, rhob]);
nothing #hide

Just fill mod with average (reference) values.

fill!(mod);
nothing #hide

Once the basic medium parameters are input, we can access some other derived parameters.

mod[:Zp];
nothing #hide

Otherwise, we can manually update parameters of mod.

mod[:vp].=3000.;
mod[:vs].=2000.;
nothing #hide

A model can be also be updated by adding random noise.

update!(mod, [:vp,:rho], randn_perc=1.);
nothing #hide

Methods

Base.getindex — Method

Get other dependent model parameters of a seismic model that are not present in Seismic.

:rhoI : inverse of density

:Zp : P-wave impedance

GeoPhyInv.update! — Method

In-place method to add features to the input model.

Arguments

mod::Medium : model that is modified

Keyword Arguments

point_loc::Vector{Float64}=[0., 0.,] : approx location of point pert

point_pert::Float64=0.0 : perturbation at the point scatterer

ellip_loc::Vector{Float64}=nothing : location of center of perturbation, [z, x]

ellip_rad::Float64=0.0 : size of elliptic perturbation

ellip_pert::Float64=0.1 : perturbation inside the ellipse

ellip_α=0.0 : rotate the ellipse

rect_loc::Array{Float64}=nothing : rectangle location, [zmin, xmin, zmax, xmax]

rect_pert::Float64=0.1 : perturbation in a rectangle

constant_pert::Float64=0.0 : constant perturbation

randn_pert::Float64=0.0 : percentage of reference values for additive random noise

fields::Vector{Symbol}=[:χvp,:χrho,:χvs] : which fields are to be modified?

onlyin : mod is modified only when field values are in these ranges

Base.copyto! — Method

Copy medium parameters from Medium mod, to vector x. Return x. copyto!(x,mod,[:vp])

Base.vec — Method

Return a vector of medium parameters x, from Medium mod. x=vec(mod)